requirements.txt
python-telegram-bot>=20.0
pyatv
yt-dlp
ytmusicapi
python-dotenv
aiohttp
